+++ src/Components/Header/index.js
@@ -53,6 +53,7 @@ class Header extends Component {
   handleLogout() {
     localStorage.removeItem('userId');
     localStorage.removeItem('username');
+    localStorage.removeItem('classId');
     this.setState({ redirect: '/', loggedIn: false });
   }
